+++ ChangeLog.md
@@ -1,3 +1,8 @@
 ## 0.0.2
 
 * Added support for cell comments
+
+## 0.0.3
+
+* pck_enerko_reports2.p_evaluate_workbook is now public api
+* Added support for reports that are based on tables of t_er_comment_definitions (see pck_enerko_reports2.f_create_report_from_dataset)